# 60% perf improvement on large objects