--- a/package.json
+++ b/package.json
@@ -40,10 +40,10 @@
     "chalk": "^1.1.1",
     "core-util-is": "^1.0.2",
     "fast-json-parse": "^1.0.0",
-    "fast-safe-stringify": "^1.0.4",
+    "fast-safe-stringify": "^1.0.8",
     "flatstr": "^1.0.0",
     "once": "^1.3.3",
-    "quick-format": "^2.0.0",
+    "quick-format": "^2.0.3",
     "split2": "^2.0.1"
   }
 }