@davidmarkclements 2016-04-07
as an aside, a really really big object **could** blow the call stack. 
but I'm okay with this, that's a good way of saying - the object you're trying to log... it's too big